🔀 Merge: Feat#1-Entity <-> Backend

See merge request s07-blockchain-nft-sub2/S07P22A607!4

+++ backend/src/main/java/com/trippiece/backend/api/domain/repository/DecorationRepository.java
@@ -2,5 +2,5 @@ package com.trippiece.backend.api.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DecorationRepository extends JpaRepository<DecorationRepository,Integer> {
+public interface DecorationRepository extends JpaRepository<DecorationRepository, Long> {
 }
