+++ backend/src/main/java/com/trippiece/backend/api/domain/repository/DiaryRepository.java
@@ -2,5 +2,5 @@ package com.trippiece.backend.api.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DiaryRepository extends JpaRepository<DiaryRepository,Integer> {
+public interface DiaryRepository extends JpaRepository<DiaryRepository, Long> {
 }
