+++ backend/src/main/java/com/trippiece/backend/api/domain/repository/FrameRepository.java
@@ -2,5 +2,5 @@ package com.trippiece.backend.api.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FrameRepository extends JpaRepository<FrameRepository,Integer> {
+public interface FrameRepository extends JpaRepository<FrameRepository, Long> {
 }
