+++ backend/src/main/java/com/trippiece/backend/api/domain/repository/QRlogRepository.java
@@ -2,5 +2,5 @@ package com.trippiece.backend.api.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface QRlogRepository extends JpaRepository<QRlogRepository, Integer> {
+public interface QRlogRepository extends JpaRepository<QRlogRepository, Long> {
 }
