+++ backend/src/main/java/com/trippiece/backend/api/domain/repository/RegionRepository.java
@@ -2,5 +2,5 @@ package com.trippiece.backend.api.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RegionRepository extends JpaRepository<RegionRepository, Integer> {
+public interface RegionRepository extends JpaRepository<RegionRepository, Long> {
 }
