+++ backend/src/main/java/com/trippiece/backend/api/domain/repository/StickerRepository.java
@@ -2,5 +2,5 @@ package com.trippiece.backend.api.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StickerRepository extends JpaRepository<StickerRepository, Integer> {
+public interface StickerRepository extends JpaRepository<StickerRepository, Long> {
 }
