+++ backend/src/main/java/com/trippiece/backend/api/domain/repository/TripRepository.java
@@ -1,5 +1,5 @@
 package com.trippiece.backend.api.domain.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TripRepository extends JpaRepository<TripRepository, Integer> {
+public interface TripRepository extends JpaRepository<TripRepository, Long> {
 }
